feat(B-674): rename cleanup_service in data_retention_service
